Refactor stopwatchModel out of main code body
